Breaking CLI into different files
