+++ panoptes/__init__.py
@@ -5,6 +5,7 @@ organize the project properly.
 """
 
 from panoptes import aws
+from panoptes import cli
 from panoptes import generic
 from panoptes import panoptesctl
