+++ panoptes/cli/__init__.py
@@ -1,0 +1,10 @@
+""" Panoptes - CLI
+Responsible for organizing the CLI created by Click
+"""
+
+from panoptes.cli import aws
+from panoptes.cli import gcp
+
+
+if __name__ == "__main__":
+    pass
