+++ panoptes/cli/aws/__init__.py
@@ -1,0 +1,8 @@
+""" Panoptes - CLI - AWS
+Responsible for organizing commands from AWS CLI
+"""
+
+from panoptes.cli.aws import commands
+
+if __name__ == "__main__":
+    pass
