+++ panoptes/cli/gcp/__init__.py
@@ -1,0 +1,8 @@
+""" Panoptes - CLI - GCP
+Responsible for organizing commands from GCP CLI
+"""
+
+from panoptes.cli.gcp import commands
+
+if __name__ == "__main__":
+    pass
